Move CompareImages to SEUtils
