+++ SEUtils/SEUtils/CompareImages.h
